Fix hosted wallet unit test imports
